Added in some throws and exception handling
out_of_range exception thrown when trying to access outside of matrix.

Some exception handling for when heap can't be allocated when trying to make the matrix.
If bad_alloc caught, then send an error message that it was caught.
Also make row or columns to 0 and keep matrix uniform, especially if caught when allocating each row.

Signed-off-by: FDIL <fdil.fadilh1@gmail.com>